+++ Libraries/LibWeb/CMakeLists.txt
@@ -755,6 +755,7 @@ set(SOURCES
     Streams/Transformer.cpp
     Streams/TransformStream.cpp
     Streams/TransformStreamDefaultController.cpp
+    Streams/TransformStreamOperations.cpp
     Streams/UnderlyingSink.cpp
     Streams/UnderlyingSource.cpp
     Streams/WritableStream.cpp
