+++ Libraries/LibWeb/Encoding/TextEncoderStream.cpp
@@ -12,8 +12,8 @@
 #include <LibWeb/Bindings/Intrinsics.h>
 #include <LibWeb/Bindings/TextEncoderStreamPrototype.h>
 #include <LibWeb/Encoding/TextEncoderStream.h>
-#include <LibWeb/Streams/AbstractOperations.h>
 #include <LibWeb/Streams/TransformStream.h>
+#include <LibWeb/Streams/TransformStreamOperations.h>
 #include <LibWeb/WebIDL/Promise.h>
 
 namespace Web::Encoding {
